header_css

+++ src/app/employee-by-id/employee-by-id.component.ts
@@ -7,10 +7,9 @@ import {Location} from '@angular/common';
 import { Router } from '@angular/router';
 import {BetweenComponentsService} from './../between-components.service';
 
-//modal
+/*modal
 import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
-import { ModalComponent } from '../modal/modal.component';
-import { MatDialogRef } from '@angular/material/dialog';
+import { ModalComponent } from './modal/modal.component';*/
 
 @Component({
   selector: 'app-employee-by-id',
@@ -27,7 +26,7 @@ export class EmployeeByIdComponent implements OnInit {
   idEmployee;
   selectedEmployee;
 
-  constructor(public dialogRef: MatDialogRef<ModalComponent>, public matDialog: MatDialog, private employeesService : EmployeesService, private betweenComponents: BetweenComponentsService, private _Activatedroute:ActivatedRoute,private router: Router, private _location: Location) { }
+  constructor(private employeesService : EmployeesService, private betweenComponents: BetweenComponentsService, private _Activatedroute:ActivatedRoute,private router: Router, private _location: Location) { }
 
   async ngOnInit(): Promise<void> {
 
@@ -47,13 +46,18 @@ export class EmployeeByIdComponent implements OnInit {
 //_______________________________________
 
 toDelete(){
-   this.employeesService.deleteEmployee(this.idEmployee).subscribe(result =>{
-    console.log("todelete: " + JSON.stringify(result))
-   });
-   setTimeout(() => {
-    this.router.navigate(['/employees']);
-}, 500);
+  let bool;
+    bool = confirm("delete?");
+    console.log(bool);
+    if(bool == true){
+      this.employeesService.deleteEmployee(this.idEmployee).subscribe(result =>{
+      console.log("todelete: " + JSON.stringify(result))
+      });
+      setTimeout(() => {
+      this.router.navigate(['/employees']);
+      }, 500);
   }
+}
 
 //_______________________________________
   toUpdate(): void {
@@ -68,34 +72,5 @@ toDelete(){
   goBack(){
     this._location.back();
   }
-
-//__________________________________________________
-//modal
-/*
-openModal() {
-  const dialogConfig = new MatDialogConfig();
-  // The user can't close the dialog by clicking outside its body
-  dialogConfig.disableClose = true;
-  dialogConfig.id = "modal-component";
-  dialogConfig.height = "350px";
-  dialogConfig.width = "600px";
-  // https://material.angular.io/components/dialog/overview
-  const modalDialog = this.matDialog.open(ModalComponent, dialogConfig);
-}
-
-
-  actionFunction() {
-    alert("You have logged out.");
-    this.closeModal();
-  }
-
-  // If the user clicks the cancel button a.k.a. the go back button, then\
-  // just close the modal
-  closeModal() {
-    this.dialogRef.close();
-  }
-
-*/
-
 
 }//closes classe
